# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: WorkLink API
  version: 0.35.0
  description: API V1 do WorkLink para descoberta, contato, avaliações, denúncias, administração e métricas.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/authentication/register:
    post:
      summary: Cria conta local e inicia uma sessão.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalRegistrationRequest"
      responses:
        "201":
          description: Conta e sessão criadas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationSession"
        "400":
          description: Cadastro inválido ou não concluído.
  /api/v1/authentication/login:
    post:
      summary: Autentica uma conta local por email e senha.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalLoginRequest"
      responses:
        "200":
          description: Sessão iniciada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationSession"
        "401":
          description: Credenciais inválidas ou conta temporariamente bloqueada.
  /api/v1/authentication/password-recovery/request:
    post:
      summary: Solicita recuperação sem revelar se o email existe.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordRecoveryRequest"
      responses:
        "202":
          description: Solicitação processada com resposta genérica.
  /api/v1/authentication/password-recovery/reset:
    post:
      summary: Redefine a senha com token de uso único.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "204":
          description: Senha redefinida e sessões anteriores revogadas.
        "400":
          description: Token ou nova senha inválidos.
  /api/v1/authentication/session/revoke:
    post:
      summary: Revoga a sessão correspondente ao refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "204":
          description: Sessão revogada.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    LocalRegistrationRequest:
      type: object
      required:
        - fullName
        - phoneNumber
        - emailAddress
        - password
        - passwordConfirmation
        - legalAccepted
      properties:
        fullName:
          type: string
        phoneNumber:
          type: string
        emailAddress:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
        passwordConfirmation:
          type: string
          format: password
        legalAccepted:
          type: boolean
    LocalLoginRequest:
      type: object
      required:
        - emailAddress
        - password
      properties:
        emailAddress:
          type: string
          format: email
        password:
          type: string
          format: password
    PasswordRecoveryRequest:
      type: object
      required:
        - emailAddress
      properties:
        emailAddress:
          type: string
          format: email
    PasswordResetRequest:
      type: object
      required:
        - recoveryToken
        - newPassword
        - newPasswordConfirmation
      properties:
        recoveryToken:
          type: string
        newPassword:
          type: string
          format: password
          minLength: 12
          maxLength: 128
        newPasswordConfirmation:
          type: string
          format: password
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthenticationSession:
      type: object
      required:
        - customerIdentifier
        - accessToken
        - refreshToken
        - accessTokenExpiresAt
        - refreshTokenExpiresAt
      properties:
        customerIdentifier:
          type: string
          format: uuid
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        refreshTokenExpiresAt:
          type: string
          format: date-time